requirements.txt
streamlit
python-dotenv
google-generativeai
Pillow
pdf2image
PyPDF2 